[otbn] Refactor test parameter generation script
We have a script to produce parameters to be used in the crypto tests
for OTBN. Refactor this script to be extensible to algorithms other than
RSA. Also align the output with the code we have in the RSA test today
to make copy-pasting easier.
